Fixed: Deletion of a customer could delete all newsletter subscriptions

+++ src/Presentation/SmartStore.Web/Administration/Controllers/CustomerController.cs
@@ -1038,30 +1038,33 @@ namespace SmartStore.Admin.Controllers
 
             var customer = _customerService.GetCustomerById(id);
             if (customer == null)
-                //No customer found with the specified id
                 return RedirectToAction("List");
 
             try
             {
                 _customerService.DeleteCustomer(customer);
 
-				//remove newsletter subscriptions (if exists)
-				var subscriptions = _newsLetterSubscriptionService.GetAllNewsLetterSubscriptions(customer.Email, 0, int.MaxValue, true);
-
-				foreach (var subscription in subscriptions)
+				if (customer.Email.HasValue())
 				{
-					_newsLetterSubscriptionService.DeleteNewsLetterSubscription(subscription);
+					foreach (var store in _storeService.GetAllStores())
+					{
+						var subscription = _newsLetterSubscriptionService.GetNewsLetterSubscriptionByEmail(customer.Email, store.Id);
+						if (subscription != null)
+						{
+							_newsLetterSubscriptionService.DeleteNewsLetterSubscription(subscription);
+						}
+					}
 				}
 
-                //activity log
-                _customerActivityService.InsertActivity("DeleteCustomer", _localizationService.GetResource("ActivityLog.DeleteCustomer"), customer.Id);
+                _customerActivityService.InsertActivity("DeleteCustomer", T("ActivityLog.DeleteCustomer", customer.Id));
 
-                NotifySuccess(_localizationService.GetResource("Admin.Customers.Customers.Deleted"));
+                NotifySuccess(T("Admin.Customers.Customers.Deleted"));
+
                 return RedirectToAction("List");
             }
-            catch (Exception exc)
+            catch (Exception exception)
             {
-				NotifyError(exc.Message);
+				NotifyError(exception.Message);
                 return RedirectToAction("Edit", new { id = customer.Id });
             }
         }
